Cleaning up ACER a little bit.
