+++ src/lipservice/ACER/Card4.hpp
@@ -7,14 +7,13 @@ public:
   template< typename Char >
   Card4( iRecordStream< Char >& is,
          const Argument< ACER::Card2::Nxtra >& nxtra )
-    try:
-      izaw( argument::extract< Izaw >( is, nxtra ) )
-    {
-      Card::clear( is );
-    }
-    catch( std::exception& e ){
-      Log::info("Trouble while validating Card 4");
-      throw e;
-    }
+  try:
+    izaw( argument::extract< Izaw >( is, nxtra ) )
+  {
+    Card::clear( is );
+  } catch( std::exception& e ){
+    Log::info("Trouble while validating Card 4");
+    throw e;
+  }
 };
