+++ src/lipservice/ACER/Card8a.hpp
@@ -6,9 +6,9 @@ public:
   Argument< Iza > iza;
 
   template< typename Char >
-  Card8a( iRecordStream< Char >& is )
+  Card8a( iRecordStream< Char >& is, const int nza )
   try:
-    iza( argument::extract< Iza >( is ) )
+    iza( argument::extract< Iza >( is, nza ) )
   {
     Card::clear( is );
   } catch( std::exception& e ){
